Aca les subo el codigo completo y con algunos arreglos de lineas que no se usaban

+++ src/tp2/laboratorio/pkg1/perry/OrniAzul.java
@@ -1,11 +1,9 @@
 package tp2.laboratorio.pkg1.perry;
 // Acá cree una parte del OrniAzul - en breve agregamos los métodos
 
-import java.util.Scanner;
-
 
 public class OrniAzul extends PapaCastor{
- private String nombre;
+    private String nombre;
     private double propulsion;
     private double pico = 6.8; // cm
     private double patas = 7.5; // cm
@@ -18,7 +16,7 @@ public class OrniAzul extends PapaCastor{
             this.propulsion = propulsion;
         } else {
             System.out.println("Advertencia: Propulsión fuera de rango (5-10 km/s). Asignando valor por defecto: 5 km/s");
-            this.propulsion = (int) 5.0;
+            this.propulsion =  5.0;
         }
     }
      @Override
@@ -37,8 +35,8 @@ public class OrniAzul extends PapaCastor{
         return propulsion;
     }
     
-    public int getVelocidadTotal() {
-         return (int) (super.getVelocidad() + propulsion);
+    public double getVelocidadTotal() {
+         return super.getVelocidad() + propulsion;
     
     }
 }
